+++ include/raknet/DR_SHA1.h
@@ -283,7 +283,7 @@ public:
 
 unsigned char * GetHash( void ) const;
 // KevinJ: http://cseweb.ucsd.edu/~mihir/papers/hmac-cb.pdf
-    static void HMAC(unsigned char *sharedKey, int sharedKeyLength, unsigned char *data, int dataLength, unsigned char output[SHA1_LENGTH]);
+    static void HMAC(unsigned char *sharedKey, size_t sharedKeyLength, unsigned char *data, size_t dataLength, unsigned char output[SHA1_LENGTH]);
 
 private:
     // Private SHA-1 transformation
