+++ include/raknet/FileOperations.h
@@ -12,6 +12,7 @@
 ///
 
 
+#include <cstddef>
 #include "NativeFeatureIncludes.h"
 #if _RAKNET_SUPPORT_FileOperations==1
 
@@ -20,7 +21,7 @@
 
 #include "Export.h"
 
-bool RAK_DLL_EXPORT WriteFileWithDirectories( const char *path, char *data, unsigned dataLength );
+bool RAK_DLL_EXPORT WriteFileWithDirectories( const char *path, char *data, size_t dataLength );
 bool RAK_DLL_EXPORT IsSlash(unsigned char c);
 void RAK_DLL_EXPORT AddSlash( char *input );
 void RAK_DLL_EXPORT QuoteIfSpaces(char *str);
